% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcereal.R
\name{last_version}
\alias{last_version}
\title{Return the latest version of cereal on the GitHub.}
\usage{
last_version()
}
\value{
\code{package_version}
}
\description{
Uses the GitHub API to find the latest version of cereal.
}
\details{
Gets all the versions from GitHub via \code{\link[=list_version]{list_version()}} and selects the
largest version number.
}
