% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcereal.R
\name{list_version}
\alias{list_version}
\title{List version(s) of cereal on GitHub.}
\usage{
list_version()
}
\value{
\code{package_version}, a vector of available versions.
}
\description{
Use the GitHub API to query the versions of cereal.
}
\details{
The GitHub page of cereal is \url{https://github.com/USCiLab/cereal}, the
tags are accessed via the GitHub API, from which a \code{package_version} object
is coerced.
}
