% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcereal.R
\name{update_version}
\alias{update_version}
\title{Update installed cereal headers}
\usage{
update_version(version = last_version(), ...)
}
\arguments{
\item{version}{\code{character} or \link{package_version}; the version to install, e.g.
'1.3.2' or \code{v1.3.2}.}

\item{...}{additional arguments passed to \code{\link[=system.file]{system.file()}}, e.g. \code{lib.loc} for
the location of the library that Rcereal is installed in.}
}
\description{
Clone a different version of the cereal headers into R library.
}
\details{
This over-writes the installed cereal headers inside an R library. The
default location for the files is found via \code{\link[=system.file]{system.file()}}. The library
location can be specified by passing an argument \code{lib.loc}. See
\code{\link[=system.file]{system.file()}} for further details.
}
