% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rcereal.R
\name{last_version}
\alias{last_version}
\title{Return the latest version of cereal on the github.}
\usage{
last_version()
}
\value{
package_version. The latest version of cereal.
}
\description{
Return the latest version of cereal on the github.
}
\details{
This function requires the package \code{httr}.
}

