% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rcereal.R
\name{list_version}
\alias{list_version}
\title{List the version of cereal on github.}
\usage{
list_version()
}
\value{
package_version. The vector of versions of cereal.
}
\description{
List the version of cereal on github.
}
\details{
Use the github API to query the versions of cereal.
 The github page of cereal is \url{https://github.com/USCiLab/cereal}
 This function requires the package \code{httr}.
}

