% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rcereal.R
\name{update_version}
\alias{update_version}
\title{Update the cereal to the Given Version}
\usage{
update_version(version = last_version())
}
\arguments{
\item{version}{package_version. The version of cereal
you want to switch in Rcereal.}
}
\description{
Update the cereal to the Given Version
}
\details{
This function will change the version of cereal
 in Rcereal. The package \code{git2r} is required.
}

